Move default limits to constants.ts, move some files from src/common to src/webgpu
