+++ src/webgpu/api/validation/error_scope.spec.ts
@@ -3,9 +3,9 @@ error scope validation tests.
 `;
 
 import { Fixture } from '../../../common/framework/fixture.js';
-import { getGPU } from '../../../common/framework/gpu/implementation.js';
 import { makeTestGroup } from '../../../common/framework/test_group.js';
 import { assert, raceWithRejectOnTimeout } from '../../../common/framework/util/util.js';
+import { getGPU } from '../../util/navigator_gpu.js';
 
 class F extends Fixture {
   _device: GPUDevice | undefined = undefined;
