+++ src/webgpu/util/navigator_gpu.ts
@@ -1,6 +1,6 @@
 /// <reference types="@webgpu/types" />
 
-import { assert } from '../util/util.js';
+import { assert } from '../../common/framework/util/util.js';
 
 let impl: GPU | undefined = undefined;
